+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 current
 =======
 
+* Replace `string-max-length` by `max-message-length`/`-M`. This is not compatible with version 0.4.4.
+
+
 0.4.4
 ======